refactor and updated for latest release

+++ scripts/deploy.sh
@@ -2,12 +2,26 @@
 cd deploy_goLogScythe || { echo "💥 💥 directory : deploy_goLogScythe  not found, will exit"; exit 1; }
 TODAY=$(date -I)
 echo "${TODAY}"
-mkdir /var/lib/go-log-scythe/
-mkdir /etc/go-log-scythe/
-rsync -av config.env /etc/go-log-scythe/
-rsync -av go-log-scythe.service /etc/systemd/system/go-log-scythe.service
-rsync -av nftables.conf /etc/
-rsync -av go-log-scythe /var/lib/
+echo "## will create directories, if already exists, will jump to next"
+if [ ! -d "/var/lib/go-log-scythe/" ]; then
+    mkdir /var/lib/go-log-scythe/
+fi
+if [ ! -d "/etc/go-log-scythe/" ]; then
+    mkdir /etc/go-log-scythe/
+fi
+echo "## will copy files if not already exists"
+if [ ! -f "/etc/go-log-scythe/config.env" ]; then
+    rsync -av config.env /etc/go-log-scythe/
+fi
+if [ ! -f "/etc/systemd/system/go-log-scythe.service" ]; then
+    rsync -av go-log-scythe.service /etc/systemd/system/go-log-scythe.service
+fi
+if [ ! -f "/etc/nftables.conf" ]; then
+    rsync -av nftables.conf /etc/
+fi
+if [ ! -f "/var/lib/go-log-scythe/go-log-scythe" ]; then
+    rsync -av go-log-scythe /var/lib/
+fi
 ufw status
 crontab -l
 ufw status > "ufw_status_${TODAY}.txt"
@@ -36,7 +50,7 @@ journalctl --disk-usage
 echo "## adjust journalctl allowed disk usage (SystemMaxUse=1G)"
 vim /etc/systemd/journald.conf
 echo "## retrieving goLogScythe release"
-wget https://github.com/lao-tseu-is-alive/go-log-scythe/releases/download/v0.2.1/goLogScythe-linux-amd64.tar.gz
+wget https://github.com/lao-tseu-is-alive/go-log-scythe/releases/download/v0.4.0/goLogScythe-linux-amd64.tar.gz
 tar xvfz goLogScythe-linux-amd64.tar.gz
 mv goLogScythe-linux-amd64 /usr/local/bin/
 systemctl daemon-reload
